split tests in new file
